LayoutTests/editing/mac/selection/context-menu-select-editability-expected.txt

This test checks that conext menu selection allows whitespace for non-editable fields. To test manually, right click on the blank text in the input box.

The crowd says: New York, New York    is a helluva town!
The crowd says: New York, New York    is a helluva town!




PASS getSelectionForId('inputWhitespace') is ""
PASS getSelectionForId('readOnlyWhitespace') is ""
PASS successfullyParsed is true

TEST COMPLETE
